+++ lib/src/core/func.dart
@@ -1738,12 +1738,15 @@ class Func1<T, R> {
     required List<SagaStep<dynamic, dynamic>> steps,
     void Function(int index, dynamic result)? onCompensate,
     void Function(int index, dynamic result)? onStepComplete,
+    void Function(int index, dynamic result, Object error)?
+        onCompensationError,
   }) {
     return SagaExtension1(
       this,
       steps: steps,
       onCompensate: onCompensate,
       onStepComplete: onStepComplete,
+      onCompensationError: onCompensationError,
     );
   }
 
@@ -2647,12 +2650,15 @@ class Func2<T1, T2, R> {
     required List<SagaStep<dynamic, dynamic>> steps,
     void Function(int index, dynamic result)? onCompensate,
     void Function(int index, dynamic result)? onStepComplete,
+    void Function(int index, dynamic result, Object error)?
+        onCompensationError,
   }) {
     return SagaExtension2(
       this,
       steps: steps,
       onCompensate: onCompensate,
       onStepComplete: onStepComplete,
+      onCompensationError: onCompensationError,
     );
   }